Selection updates: Sync selection of objects and bodies.
This avoids the problem of editing unrelated data and makes tweaking body settings more intuitive. The mechanism ensures that the selected object in the scene is always one that is used by the active body, either by changing the object when another body is selected, or by setting the active body if the object does not match it.